requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
jinja2
python-dotenv
